requirements.txt
Flask==2.3.3
gunicorn==21.2.0
openpyxl==3.1.2
psycopg2-binary